añado test unitarios y elimino dataset

+++ backendfile/appFile.py
@@ -1,3 +1,10 @@
+#   Trabajo Fin de Máster
+#   Máster en Ingeniería Informática
+#
+#   2020 - Copyright (c) - GNU v3.0
+#
+#  Matilde Cabrera <mati331@correo.ugr.es>
+
 from flask import Flask, request
 from flask_cors import CORS
 import json
@@ -56,16 +63,16 @@ Typefiles = {
     'filesTIS'}
 
 # return 'hello world', test metod
-@app.route('/')
-@cache.cached(timeout=50)
-def hello_world():
-    return jsonify('Hello, World !'), 200
+#@app.route('/')
+#@cache.cached(timeout=50)
+#def hello_world():
+#    return jsonify('Hello, World !'), 200
 
 # return all of mongodb, test metod
-@app.route('/todo', methods = ['GET','POST'])
-@cache.cached(timeout=50)
-def todo():
-    return  jsonify(client.getAll()), 200
+#@app.route('/todo', methods = ['GET','POST'])
+#@cache.cached(timeout=50)
+#def todo():
+#    return  jsonify(client.getAll()), 200
 
 # new client
 @app.route('/newclient/<string:correo_id>', methods = ['GET','POST'])
@@ -84,20 +91,22 @@ def deleteClient(correo_id):
 @app.route('/files/<string:correo_id>', methods=['GET'])
 def filesClient(correo_id):
     files = client.getFiles(correo_id)
-    for i in files:
-        filesFVS = i['filesFVS']
-        filesFDS = i['filesFDS']
-        filesTIS = i['filesTIS'] 
-        files = {'filesFVS':filesFVS, 'filesFDS': filesFDS, 'filesTIS': filesTIS}
-        return jsonify(files), 200
-    return jsonify("error"), 500
+    print("files")
+    print(files)
+    return jsonify(files), 200
     
 @app.route('/removefile/<string:correo_id>/<string:typefile>', methods = ['DELETE'])
 def removefile(correo_id,typefile):
+    if (client.getCliente(correo_id) is False):
+        return "Usuario NO existe, operación no valida"
     shutil.rmtree(app.config['UPLOAD_FOLDER']+'/'+correo_id+'/'+ typefile)
     os.mkdir(app.config['UPLOAD_FOLDER']+'/'+correo_id+'/'+typefile)
     files = client.deleteFiles(correo_id,typefile)
-    return jsonify('ok'), 200
+    if files:
+        salida = "archivos eliminados", 200
+    else:
+        salida = "error, los archivos no han sido eliminados", 400
+    return jsonify(salida)
 
 # insert file in folder clients
 @app.route('/upload', methods = ['POST'])
@@ -108,7 +117,7 @@ def upload():
         return 'No has seleccionado el tipo de archivo', 400
     typefile = request.form['typefile']
     if typefile not in Typefiles:
-        return "No existe el tipo de archivo seleccionado "
+        return "No existe el tipo de archivo seleccionado ",400
 
     if 'email' not in request.form:
         return 'Es necesario que se identifique', 400
@@ -136,6 +145,9 @@ def upload():
         # Guardo nombres en mongodb
         response = client.insertFiles(correo_id,typefile,filenamepos, filenameneg)
 
+        if response == "Usuario NO existe, operación no valida":
+            return jsonify(response), 500
+
         # el directorio para el usuario actual
         myfolder = app.config['UPLOAD_FOLDER']+'/'+correo_id
         # si no existe el directorio lo creamos
@@ -153,21 +165,30 @@ def upload():
 def search():
     response = ''
     body = request.get_json() # obtener el contenido del cuerpo de la solicitud
+    
     if body is None:
-        return "The request body is null", 400
-    if body['positive'] is None:
-        return 'You need to specify the file positive',400
-    if body['negative'] is None:
-        return 'You need to specify the file negative',400
+        return "Request body es null", 400
+    if body['positive'] == '':
+        return 'Necesita especificar el archivo  positive',400
+    if body['negative'] == '':
+        return 'Necesita especificar el archivo negative',400
     if body['typefile'] == '':
-        return 'You need to specify the typefile',400
+        return 'Necesita especificar el tipo de archivo',400
     if body['wanted'] == '':
-        return 'You need to specify what looking for',400
+        return 'Necesita especificar que busca',400
+    if body['email'] == '':
+        return 'Es necesario que se identifique',400
 
     filepos = body['positive']
     fileneg = body['negative']
     typefile = body['typefile']
     correo = body['email']
+
+    if (client.getCliente(correo) is False):
+        return "Usuario NO existe, operación no valida"
+    if typefile not in Typefiles:
+        return "No existe el tipo de archivo seleccionado ",400
+
     count = filepos.find(".txt")
     count2 = fileneg.find(".txt")
     pathfile = app.config['UPLOAD_FOLDER']+'/'+correo+'/'+typefile+'/lucene_'+filepos[:count]
@@ -179,21 +200,29 @@ def search():
         resultado.append(i)
     return jsonify(resultado), 200
 
+# Obtener los términos frecuencia de un par de archivos seleccionados.
 @app.route('/tis', methods = ['GET', 'POST'])
 def selectTIS():
     response = ''
     body = request.get_json() # obtener el contenido del cuerpo de la solicitud
     if body is None:
-        return "The request body is null", 400
-    if body['positive'] is None:
-        return 'You need to specify the file positive',400
-    if body['negative'] is None:
-        return 'You need to specify the file negative',400
+        return "Request body es null", 400
+    if body['positive'] == '':
+        return 'Necesita especificar el archivo  positive',400
+    if body['negative'] == '':
+        return 'Necesita especificar el archivo negative',400
+    if body['typefile'] != 'filesTIS':
+        return 'Sólo se admiten archivos TIS',400
+    if body['email'] == '':
+        return 'Es necesario que se identifique',400
 
     filepos = body['positive']
     fileneg = body['negative']
     typefile = body['typefile']
     correo = body['email']
+    if (client.getCliente(correo) is False):
+        return "Usuario NO existe, operación no valida"
+        
     count = filepos.find(".txt")
     count2 = fileneg.find(".txt")
     pathfilepos = app.config['UPLOAD_FOLDER']+'/'+correo+'/filesTIS/lucene_'+filepos[:count]
@@ -209,12 +238,25 @@ def selectTIS():
     resultado = {'terms_freqs_positive': resultadopos, 'terms_freqs_negative': resultadoneg}
     return jsonify(resultado), 200
 
+# se obtiene una consulta
 @app.route('/filter', methods = ['GET', 'POST'])
 def filter():
     body = request.get_json()
     typefilter = body['typefilter']
     if typefilter not in Filters:
         return "No existe el filtro seleccionado "
+    if body['terms_freqs_positive'] == '':
+        return 'Necesita especificar los terminos positivos',400
+    if body['terms_freqs_negative'] == '':
+        return 'Necesita especificar los terminos negativos',400
+    if body['email'] == '':
+        return 'Es necesario que se identifique',400
+    if body['sum'] == '':
+        return 'No ha especificado cuantas palabras quiere',400
+
+    if (client.getCliente(body['email']) is False):
+        return "Usuario NO existe, operación no valida"
+
     terms_freqs_positive = body['terms_freqs_positive']
     print(terms_freqs_positive)
     terms_freqs_negative = body['terms_freqs_negative']
@@ -228,6 +270,7 @@ def filter():
     # devuelvo la consulta ???
     return jsonify(resultado), 200
 
+# obtiene las medidas de rendimiento
 @app.route('/applyFilter', methods = ['GET', 'POST'])
 def applyFilter():
     response = ''
@@ -235,23 +278,23 @@ def applyFilter():
     if body is None:
         return "The request body is null", 400
     if body['email'] is None:
-        return 'You need to identify yourself', 400
+        return 'Es necesario que se identifique', 400
     if body['wanted'] == '':
-        return 'You need to specify what looking for',400
+        return 'Necesita especificar que busca',400
 
     correo = body['email']
+    if (client.getCliente(body['email']) is False):
+        return "Usuario NO existe, operación no valida"
     files = client.getFiles(correo)
-    for i in files:
-        filesFVS = i['filesFVS']
-        filesFDS = i['filesFDS']
-        filesTIS = i['filesTIS'] 
-        files = {'filesFVS':filesFVS, 'filesFDS': filesFDS, 'filesTIS': filesTIS}
 
     salida = {}
     if body['typefile'] == 'filesFVS':
-        filesX = filesFVS
+        filesX = files['filesFVS']
     elif body['typefile']== 'filesFDS':
-        filesX = filesFDS
+        filesX = files['filesFDS']
+    else:
+        return "Los tipos de archivos solo pueden ser FVS o FDS", 400
+
     for e in filesX:
         resultado = {}
         setAllPos = {}
@@ -288,6 +331,8 @@ def createDirectory(myfolder):
         os.mkdir(myfolder+'/filesFVS')
         os.mkdir(myfolder+'/filesFDS')
         os.mkdir(myfolder+'/filesTIS')
+
+
 
 @app.after_request
 def after_request(response):
